feat: cartUtils 테스트 완료
