+++ src/refactoring/models/cart.ts
@@ -1,21 +1,54 @@
-import { CartItem, Coupon } from "../../types";
+import { CartItem, Coupon } from '../../types';
 
 export const calculateItemTotal = (item: CartItem) => {
-  return 0;
+  const { price } = item.product;
+  const { quantity } = item;
+  const discount = getMaxApplicableDiscount(item);
+  return price * quantity * (1 - discount);
 };
 
 export const getMaxApplicableDiscount = (item: CartItem) => {
-  return 0;
+  const { quantity } = item;
+  return item.product.discounts.reduce(
+    (maxDiscount, d) =>
+      quantity >= d.quantity && d.rate > maxDiscount ? d.rate : maxDiscount,
+    0
+  );
 };
 
 export const calculateCartTotal = (
   cart: CartItem[],
   selectedCoupon: Coupon | null
 ) => {
+  const totalBeforeDiscount = cart.reduce((acc, item) => {
+    const { price } = item.product;
+    const { quantity } = item;
+    return acc + price * quantity;
+  }, 0);
+
+  let totalAfterDiscount = cart.reduce(
+    (acc, item) => acc + calculateItemTotal(item),
+    0
+  );
+
+  let totalDiscount = totalBeforeDiscount - totalAfterDiscount;
+
+  if (selectedCoupon) {
+    if (selectedCoupon.discountType === 'amount') {
+      totalAfterDiscount = Math.max(
+        0,
+        totalAfterDiscount - selectedCoupon.discountValue
+      );
+    } else {
+      totalAfterDiscount *= 1 - selectedCoupon.discountValue / 100;
+    }
+    totalDiscount = totalBeforeDiscount - totalAfterDiscount;
+  }
+
   return {
-    totalBeforeDiscount: 0,
-    totalAfterDiscount: 0,
-    totalDiscount: 0,
+    totalBeforeDiscount: Math.round(totalBeforeDiscount),
+    totalAfterDiscount: Math.round(totalAfterDiscount),
+    totalDiscount: Math.round(totalDiscount),
   };
 };
 
@@ -24,5 +57,12 @@ export const updateCartItemQuantity = (
   productId: string,
   newQuantity: number
 ): CartItem[] => {
-  return [];
+  const updatedCart = cart.map((item) => {
+    if (item.product.id === productId) {
+      return { ...item, quantity: Math.min(newQuantity, item.product.stock) };
+    }
+    return item;
+  });
+
+  return updatedCart.filter((item) => item.quantity > 0);
 };
